Fix format() to behave like in MySQL, add tests
* Input type is now numeric to match MySQL
* Always add a leading zero for numbers less than 1
* Round input number to the specified fractional places
* No trailing '.' when fractional places is 0
* Work around PostgreSQL bug (to be investigated): to_char(10,'FM99.') => '1'

+++ sql/mysqlcompat_uninstall.sql
@@ -115,7 +115,7 @@ DROP FUNCTION get_format(text, text);
 DROP FUNCTION from_unixtime(bigint, text);
 DROP FUNCTION from_unixtime(bigint);
 DROP FUNCTION from_days(integer);
-DROP FUNCTION format(double precision, integer);
+DROP FUNCTION format(numeric, integer);
 DROP FUNCTION find_in_set(text, text);
 DROP FUNCTION field(text, text, text, text, text);
 DROP FUNCTION field(text, text, text, text);
